fix(CoolMessages): prevent duplicate message add
Fixes a bug where any user could star react to a message and the bot
would unconditionally add that message to the cool messages board,
considering the fact that it would surpass the minimum reaction
count requirement.

This commit gets the bot to react with the same cool messages emoji the
first time and use that reaction as proof that it has quoted the message
and that it should not deal with it in case somebody reacts to it again.