refactor: use correct method for reactionsCount
It turns out that for each event fired, every *single* damn time,
messageReaction.hasCount() would always return false. No matter what.
Terrible documentation from JDA's side. As a result, because of the
ternary operator:

    messageReaction.hasCount() ? messageReaction.getCount() + 1 : 1

the result of `reactionsCount` would always end up holding the value of
one.

In the following changes, we use `messageReaction.retrieveUsers()` to
get a list of the people reacted, get a `Stream<User>` from that and get
its count. Much more reliable this way and it also happens to be more
readable.

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>