perf(cakeDayRole): delegate role fetching to JDA
Currently we are getting the JDA to fetch all the roles for us, just so
that we filter it ourselves, however we are overfetching information
that we do not need anyways.

Only get what we want from the JDA by calling 'Guild#getRolesByName' so
that we are only returned relevant roles from the start, and from there
just get the first role in case there is a data error and we have
multiple roles.

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>