CakeDayService.java: rewrite removeRoleFromMembers()
The method removeRoleFromMembers() is written in such a way to
batch-process all of the requests to remove the role from each member
specified - however it was written in the name of lack of defensive
programming and more specifically, without handling errors for each
member having their role removed.

Rewrite it to take care of errors for each member as they lose their
role.

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>